+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Support objectOverrides using `.spec.objectOverrides`.
   See [objectOverrides concepts page](https://docs.stackable.tech/home/nightly/concepts/overrides/#object-overrides) for details ([#831]).
+- Enable the [restart-controller](https://docs.stackable.tech/home/nightly/commons-operator/restarter/), so that the Pods are automatically restarted on config changes ([#833]).
 - Add `enabledColumnMasking` field to `opa` configuration in `authorization` ([#827]).
 - Support batched column masks in Rego rules ([#827]).
 
@@ -19,6 +20,7 @@ All notable changes to this project will be documented in this file.
   `opa.policy.batch-column-masking-uri` instead, allowing Trino to fetch multiple column masks in a single request ([#827]).
 
 [#831]: https://github.com/stackabletech/trino-operator/pull/831
+[#833]: https://github.com/stackabletech/trino-operator/pull/833
 [#827]: https://github.com/stackabletech/trino-operator/pull/827
 
 ## [25.11.0] - 2025-11-07